Creating a prebuild script and adding command.
