+++ script/build.js
@@ -1,5 +1,5 @@
 // Builds the site using Metalsmith as the top-level build runner.
-
+const path = require('path');
 const Metalsmith = require('metalsmith');
 const archive = require('metalsmith-archive');
 const assets = require('metalsmith-assets');
@@ -22,33 +22,8 @@ const watch = require('metalsmith-watch');
 const webpack = require('metalsmith-webpack');
 const webpackConfigGenerator = require('../config/webpack.config');
 const webpackDevServer = require('metalsmith-webpack-dev-server');
-const semver = require('semver');
-
-const fs = require('fs');
-const path = require('path');
 
 const sourceDir = '../content/pages';
-const minimumNodeVersion = '6.11.1';
-
-if (!(process.env.INSTALL_HOOKS === 'no')) {
-  // Make sure git pre-commit hooks are installed
-  ['pre-commit'].forEach(hook => {
-    const src = path.join(__dirname, `../hooks/${hook}`);
-    const dest = path.join(__dirname, `../.git/hooks/${hook}`);
-    if (fs.existsSync(src)) {
-      if (!fs.existsSync(dest)) {
-        // Install hooks
-        fs.linkSync(src, dest);
-      }
-    }
-  });
-}
-
-if (semver.compare(process.version, minimumNodeVersion) === -1) {
-  process.stdout.write(`Node.js version (mininum): v${minimumNodeVersion}\n`);
-  process.stdout.write(`Node.js version (installed): ${process.version}\n`);
-  process.exit(1);
-}
 
 const smith = Metalsmith(__dirname); // eslint-disable-line new-cap
